requirements.txt
imapclient
python-dotenv
mail-parser
fastapi
requests
uvicorn